filter: Restructure for testability
While working for an end-to-end test for influx-spout it became
apparent that the current way the filter is started makes it difficult
to test.

The main change here is that StartFilter now returns a Filter instance
instead of calling runtime.Goexit() and Filter now has Stop() method
which cleans up the filter. This makes it possible to shutdown a Filter
once it has been started, making it much easier to work with.

This is very similar to how the writer is already structured.

Also here:
- Simplified the "medium" filter tests, taking advantage of the new
  StartFilter/Stop structure. A dodgy 1s sleep could now be removed.
- Unexported a number of types and methods that weren't being used
  externally (and never will be).
- Added docstrings to all exported types and methods.
- Renamed filter.FilterRule to just filter.Rule to avoid stuttering.