Parallelize filtering
To help it cope with large traffic volumes, the filter now has
multiple worker goroutines. The number of workers is set by the
`workers` configuration option (previously only used for the writer
component).

A new worker type encapsulates the filtering activity, allowing
various items of state to be removed from the Filter type, making the
code clearer. Each worker has it's own connection for publishing to
NATS, removing contention on the connection instance (this should help
performance).

This change also introduces a new RuleSet type which holds a
collection of Rules and providing access to them. This cleans up the
filter and writer components and their tests significantly. This has
also made way for the removal of the slightly clumsy filter.Filtering
interface.

Closes #14.