listener: Fix data races in tests
It's only valid to call testing.T.FailNow() and friends the main
goroutine. Some listener tests used goroutines with test assertions in
them, triggering data races. It turns out these goroutines were
actually unnecessary so they've been removed.

Also fixed a problem where StartListener had returned but goroutines
owner by the listener were not ready yet. This meant that sometimes
the test would start sending data to the listener before it was ready
to accept it. A `ready` channel is now used to synchronise this
situation and avoid the race.