perfcheck: Multiple benchmark iterations & uses new benchcmp
To avoid false positives, reference and current benchmarks are now run
5 times - interleaved to minimise the effects of host system load. The
outputs from each run are concatenated together resulting in a file of
benchmark output for the reference revision & the current
revision. These are then passed to the new benchstat-based benchcheck
tool.

This is approach is much more reliable than the previous approach.