Reject metrics with timestamps that are either too old or too new
A new "max_time_delta_secs" configuration option has been added. If
the filter sees a line with a timestamp which is outside
±max_time_delta_secs around the current time it will be reject. A new
metric has been added to the filter which counts these rejected lines.

This feature adds a 3-5% performance hit to the filter ProcessBatch
benchmark, although this is largely lost when compared to the cost of
matching any realistic number of filter rules (i.e. regex matching is
expensive).

Fixes #27.