listener: Prevent large HTTP POSTs from being broken up
This fixes an issue where a large HTTP POST, or a POST which arrives
when the batch buffer is nearly full could get split, corrupting one
of the lines contained within it.

The batch buffer will now be grown if required to support any HTTP
POST that arrives. Growing the batch buffer is expensive and the
listener is arranged to minimise the chance of the buffer needing to
grow at all. Growing the buffer is a fallback.