listener: Protect batch buffer from concurrent access
The HTTP listener uses concurrent HTTP handlers which were modifying
the batch buffer without synchronisation. This resulted in data races
and corrupt output from the listener. The test added in this change
would reliably result in corrupt output before the mutex was added.

The mutex is not required or used for the UDP listener as there is
only one goroutine in that case.