Implement Stats in terms of AnonStats
Stats now simply provide names for the index based counters contained
in an embedded AnonStats. This removes the use of mutex in favour of
the sync/atomic based approach of AnonStats which is faster and avoids
unnecessary synchronization between goroutines.

CPU profiling for the filter and writer used to feature stats update
and synchronization operations but these no longer show up.

The benchmark in the stats package is now ~78% faster.