Use time.Duration for all timeout values in configuration
Configuration values which were stored in seconds are now stored as a
time.Duration. This allows more convenient time duration specification
in configuration files (e.g. "5m" instead of "300") and avoids data
conversions at the points of use, reducing the chance of bugs.

As part of the change the following configuration value renames
occurred:

- batch_max_secs -> batch_max_age
- max_time_delta_secs -> max_time_delta
- write_timeout_secs -> write_timeout