Introduce new "tags" filter rule
This allows matching against one or more exact tags on measurement
lines. The approach taken is much faster than using regexes to achieve
the same kinds of matches (80-1000% depending on number tags to match
and size of line). This approach is also much safer as matching is
independent of tag order and matches tag keys and values precisely.

This change features some nice improvements in how line parsing is
done in the filter.

Close #90.