filter: Avoid unescaping unless strictly necessary
Unescaping is expensive as it involves memory allocation. Instead of
always unescaping a line, only unescape when a rule is encountered
that requires it (and then cache that for following rule to use). This
avoids unescaping when a) there's no rules that require it or b) a
match occurs before a unescape-requiring rule is reached.

This improves all the filter benchmarks, particularly
BenchmarkProcessBatch (over 1000% faster!).