listener: Never send more than 1MB to NATS at once
The goal is for the listener batch size to never end up bigger than
what NATS will accept (1MB) but there is a small chance that a series
of large incoming chunks could cause the batch to grow beyond the
intended limit. For these cases, use the batchsplitter just in
case. batchsplitter has very low overhead when no splitting is
required.