batch: Grow by at least max UDP size
The batch buffer was only being grown when there were less than 1024
bytes remaining. This was problematic in ReadOnceFrom where there
could be up to 64KB available to read (from a UDP socket), and could
result in incomplete reads.

The batch is now grown when there's less than 64KB available in the
buffer. The batch is now always grown by at least 64KB (in case a very
small batch size is configured).